gru/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.5.0
numpy>=1.24.0
requests>=2.31.0